+++ docs/ui/components/Prerequisites/Requirement.tsx
@@ -8,7 +8,9 @@ export type RequirementProps = PropsWithChildren<{
 export function Requirement({ title, children }: RequirementProps) {
   return (
     <div className="flex-1 overflow-hidden">
-      <div className="mb-2 flex items-baseline gap-2 font-medium">{title}</div>
+      <div className="mb-2 flex items-baseline gap-2 font-medium" data-md="requirement-title">
+        {title}
+      </div>
       <div className={mergeClasses('[&_p]:ml-0 [&_pre>pre]:mt-0 [&>*:last-child]:mb-0!')}>
         {children}
       </div>