+++ packages/@expo/cli/src/run/ios/index.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk';
 import path from 'path';
 
 import type { XcodeConfiguration } from './XcodeBuild.types';
-import type { Command } from '../../../bin/cli';
+import type { Command } from '../../index';
 import { assertWithOptionsArgs, printHelp } from '../../utils/args';
 import { logCmdError } from '../../utils/errors';